fix figsize type annotation to match float default in solutionconfig

+++ ultralytics/solutions/config.py
@@ -75,7 +75,7 @@ class SolutionConfig:
     down_angle: int = 90
     kpts: list[int] = field(default_factory=lambda: [6, 8, 10])
     analytics_type: str = "line"
-    figsize: tuple[int, int] | None = (12.8, 7.2)
+    figsize: tuple[float, float] | None = (12.8, 7.2)
     blur_ratio: float = 0.5
     vision_point: tuple[int, int] = (20, 20)
     crop_dir: str = "cropped-detections"
